nodemask: Remove implicit addressof from the API
The nodemask API differs from the cpumask API because each wrapper to bitmap
operations is further wrapped by a macro which takes the address of the
nodemask objects.

This results in code which is slightly confusing to read as it doesn't follow
C's calling conventions, and prohibits the use of slightly more complicated
constructs for specifying parameters.

Drop all wrapping macros, rename the nodemask static inline functions to drop
the double underscores, and feed MAX_NUMNODES into appropriate locations.

Take the opportunity to drop a compiler workaround for node_isset() for GCC
3.3.2 which is long out of support, and implment it with a static inline.

Update all callers to use the correct indirection themselves.

No functional change.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20190625144350.6176-3-andrew.cooper3@citrix.com>