x86/msi: fix loop termination condition in pci_msi_conf_write_intercept()
The for loop that deals with MSI masking is coded as follows:

for ( pos = 0; pos < entry->msi.nvec; ++pos, ++entry )

Thus the loop termination condition is dereferencing a struct pointer that
is being incremented by the loop. However, it is clear from following code
paths in msi_capability_init() that this is unsafe as for instance, in the
case of nvec == 1, entry will point at a single struct msi_desc allocation
and thus the loop will walk beyond the bounds of the allocation before
dereferencing the memory to determine whether the loop should terminate.
Also, because the body of the loop writes via the entry pointer, this can
then lead to heap memory corruption, or indeed corruption of anything in
the direct map.

This patch simply initializes a stack variable to the value of
entry->msi.nvec before starting the loop and then uses that in the
termination condition instead.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>
Message-Id: <20190702093414.27798-1-paul.durrant@citrix.com>