xen/mm.h: add helper function to test-and-clear _PGC_allocated
The _PGC_allocated flag is set on a page when it is assigned to a domain
along with an initial reference count of 1. To clear this initial
reference count it is necessary to test-and-clear _PGC_allocated and then
only drop the reference if the test-and-clear succeeds. This is open-
coded in many places. It is also unsafe to test-and-clear _PGC_allocated
unless the caller holds an additional reference.

This patch adds a helper function, clear_assignment_reference(), to
replace all the open-coded test-and-clear/put_page occurrences and
incorporates in that an ASSERTion that an additional page reference is
held.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>
Message-Id: <20190710161733.39119-1-paul.durrant@citrix.com>