schedule: mask out XEN_RUNSTATE_UPDATE from state entry time
Using XEN_RUNSTATE_UPDATE mask during the process of copying runstate
values to guest causes runstate entry time to be eventually considered
negative on a calculation of the time delta. So the XEN_RUNSTATE_UPDATE
should be masked out during the calculation of the time spent in the
particular runstate.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>
Message-Id: <1563443655-15802-1-git-send-email-andrii.anisov@gmail.com>