x86/p2m: fix non-translated handling of iommu mappings
The current usage of need_iommu_pt_sync in p2m for non-translated
guests is wrong because it doesn't correctly handle a relaxed PV
hardware domain, that has need_sync set to false, but still need
entries to be added from calls to {set/clear}_identity_p2m_entry.

Adjust the code in guest_physmap_add_page to also check whether the
domain has an iommu instead of whether it needs syncing or not in
order to take a reference to a page to be mapped.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Message-Id: <20190722153209.73107-1-roger.pau@citrix.com>
Reviewed-by: Paul Durrant <paul.durrant@citrix.com>
Tested-by: Roman Shaposhnik <roman@zededa.com>