xen/arm32: head: Setup HTTBR in enable_mmu() and add missing isb
At the moment, HTTBR is setup in create_page_tables(). This is fine as
it is called by every CPUs.

However, such assumption may not hold in the future. To make change
easier, the HTTBR is not setup in enable_mmu().

Take the opportunity to add the missing isb() to ensure the HTTBR is
seen before the MMU is turned on.

Signed-off-by: Julien Grall <julien.grall@arm.com>

Message-Id: <20190722213958.5761-35-julien.grall@arm.com>