xen: merge temporary vcpu pinning scenarios
Today there are three scenarios which are pinning vcpus temporarily to
a single physical cpu:

- NMI/MCE injection into PV domains
- wait_event() handling
- vcpu_pin_override() handling

Each of those cases are handled independently today using their own
temporary cpumask to save the old affinity settings.

The three cases can be combined as the two latter cases will only pin
a vcpu to the physical cpu it is already running on, while
vcpu_pin_override() is allowed to fail.

So merge the three temporary pinning scenarios by only using one
cpumask and a per-vcpu bitmask for specifying which of the three
scenarios is currently active (they are allowed to nest).

Note that we don't need to call domain_update_node_affinity() as we
are only pinning for a brief period of time.

Signed-off-by: Juergen Gross <jgross@suse.com>
Message-Id: <20190723092056.15045-3-jgross@suse.com>