xen: credit2: avoid using cpumask_weight() in hot-paths
cpumask_weight() is known to be expensive. In Credit2, we use it in
load-balancing, but only for knowing how many CPUs are active in a
runqueue.

Keeping such count in an integer field of the per-runqueue data
structure we have, completely avoids the need for cpumask_weight().

While there, remove as much other uses of it as we can, even if not in
hot-paths.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Message-Id: <156439734950.9656.3257482583234913248.stgit@Palanthas>
Reviewed-by: George Dunlap <george.dunlap@citrix.com>
Reviewed-by: Andrii Anisov <andrii_anisov@epam.com>