xen/nodemask: Drop nodes_{setall, clear}() and improve the initialisers
There is no need to use runtime variable-length clearing when MAX_NUMNODES is
known to the compiler.  Drop these functions and use the initialisers instead.

numa_initmem_init() opencodes nodemask_of_node(), but its implemention is
still poor, so replace it with NODEMASK_OF() which is calculated at compile
time, and without the use of a locked set_bit() operation.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20190729121204.13559-8-andrew.cooper3@citrix.com>