xen/nodemask: Sanitise the remainder of the nodemask API
The nodemask API differs from the cpumask API because each wrapper to bitmap
operations is further wrapped by a macro which takes the address of the
nodemask objects.

This results in code which is slightly confusing to read as it doesn't follow
C's calling conventions, and prohibits the use of slightly more complicated
constructs for specifying parameters.

Drop all wrapping macros, rename the nodemask static inline functions to drop
the double underscores, and feed MAX_NUMNODES into appropriate locations.

Furthermore, the naming is inconsistent.  As we're changing all callers
anyway, rationalise all the naming to be of the form nodemask_*(), and update
the types to per Xen's latest expectations.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20190729121204.13559-10-andrew.cooper3@citrix.com>